Attempting to get these columns side by side
Can't get these two columns to work in the lifting area. Same with the search bar. I'm assuming the other two worked because they were in different kits or in a list? But I am so confused with all these divs and how they work with the grids and agh brain hurts man. So much experimentation but I'm not getting any closer after like 2 hours per stupid mistake.